Changed large text columns to use "text", change java.lang.Short to short
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@202 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/datamodel/User.java
@@ -17,7 +17,7 @@ public class User  implements java.io.Serializable {
      protected Integer userId;
      protected String name;
      protected String password;
-     protected Integer privLevel;
+     protected Short privLevel;
      private Set collectionObjDef;
 
 
@@ -79,14 +79,13 @@ public class User  implements java.io.Serializable {
     /**
      *      *            @hibernate.property
      *             column="PrivLevel"
-     *             length="5"
      *         
      */
-    public Integer getPrivLevel() {
+    public Short getPrivLevel() {
         return this.privLevel;
     }
     
-    public void setPrivLevel(Integer privLevel) {
+    public void setPrivLevel(Short privLevel) {
         this.privLevel = privLevel;
     }
